Update script.js
for line 43-46, I changed the layer dense from 50 to 3. It lead to the results that in the model, the predicted data tend to shape a horizontal line, which is less accurate to the original data, which is not good.
Thus, setting the dense as 50 is better than setting it to 3.